refactor(block-amp): move full-rig NAM models to block-amp
amp+cab captures without pedals belong in block-amp, not block-full-rig.
block-full-rig is reserved for complete rigs with effects pedals baked in.

Moved 12 NAM models: ampeg_svt_classic, dover_da50_mesa, fender_bassman_1971,
fender_deluxe_reverb_65, fender_super_reverb_1977, marshall_jmp_1_full_rig,
marshall_super_100_1966, peavey_5150_mesa_4x12, roland_jc_120b_jazz_chorus,
synergy_drect_mesa, vox_ac30, vox_ac30_1961_fawn_ef86.

Capture file paths (full_rigs/...) unchanged on disk.

Closes #208
