+++ crates/block-amp/src/nam_fender_deluxe_reverb_65.rs
@@ -1,6 +1,6 @@
 use anyhow::{anyhow, Result};
-use crate::registry::FullRigModelDefinition;
-use crate::FullRigBackendKind;
+use crate::registry::{AmpBackendKind, AmpModelDefinition};
+
 use nam::{
     build_processor_with_assets_for_layout, model_schema_for,
     processor::{NamPluginParams, DEFAULT_PLUGIN_PARAMS},
@@ -8,25 +8,27 @@ use nam::{
 use block_core::param::{enum_parameter, required_string, ModelParameterSchema, ParameterSet};
 use block_core::{AudioChannelLayout, BlockProcessor};
 
-pub const MODEL_ID: &str = "vox_ac30";
-pub const DISPLAY_NAME: &str = "AC30";
-const BRAND: &str = "vox";
+pub const MODEL_ID: &str = "fender_deluxe_reverb_65";
+pub const DISPLAY_NAME: &str = "Deluxe Reverb '65";
+const BRAND: &str = "fender";
 
 pub const NAM_PLUGIN_FIXED_PARAMS: NamPluginParams = DEFAULT_PLUGIN_PARAMS;
 
-const CAPTURE_STANDARD: &str = "full_rigs/vox_ac30/vox_ac30_cab.nam";
-const CAPTURE_CLEAN_65PRINCE: &str = "full_rigs/vox_ac30/vox_ac30_clean_65prince.nam";
+const CAPTURE_SM57_ROYER: &str = "full_rigs/fender_deluxe_reverb_65/fender_drri_clean_sm57_royer.nam";
+const CAPTURE_SM57_ROYER_ROOM: &str = "full_rigs/fender_deluxe_reverb_65/fender_drri_clean_sm57_royer_room.nam";
+const CAPTURE_ROOM: &str = "full_rigs/fender_deluxe_reverb_65/fender_drri_clean_room.nam";
 
 pub fn model_schema() -> ModelParameterSchema {
-    let mut schema = model_schema_for("full_rig", MODEL_ID, DISPLAY_NAME, false);
+    let mut schema = model_schema_for("amp", MODEL_ID, DISPLAY_NAME, false);
     schema.parameters = vec![enum_parameter(
-        "character",
-        "Character",
-        Some("Amp"),
-        Some("standard"),
+        "mic",
+        "Mic",
+        Some("Cab"),
+        Some("sm57_royer"),
         &[
-            ("standard", "Standard"),
-            ("clean_65prince", "Clean 65Prince"),
+            ("sm57_royer", "SM57 + Royer R-121"),
+            ("sm57_royer_room", "SM57 + Royer R-121 + Room"),
+            ("room", "Room Only"),
         ],
     )];
     schema
@@ -57,12 +59,13 @@ pub fn asset_summary(params: &ParameterSet) -> Result<String> {
 }
 
 fn resolve_capture_path(params: &ParameterSet) -> Result<&'static str> {
-    let character = required_string(params, "character").map_err(anyhow::Error::msg)?;
-    match character.as_str() {
-        "standard" => Ok(CAPTURE_STANDARD),
-        "clean_65prince" => Ok(CAPTURE_CLEAN_65PRINCE),
+    let mic = required_string(params, "mic").map_err(anyhow::Error::msg)?;
+    match mic.as_str() {
+        "sm57_royer" => Ok(CAPTURE_SM57_ROYER),
+        "sm57_royer_room" => Ok(CAPTURE_SM57_ROYER_ROOM),
+        "room" => Ok(CAPTURE_ROOM),
         other => Err(anyhow!(
-            "full-rig model '{}' does not support character='{}'",
+            "amp model '{}' does not support mic='{}'",
             MODEL_ID,
             other
         )),
@@ -77,11 +80,11 @@ fn build(params: &ParameterSet, sample_rate: f32, layout: AudioChannelLayout) ->
     build_processor_for_model(params, sample_rate, layout)
 }
 
-pub const MODEL_DEFINITION: FullRigModelDefinition = FullRigModelDefinition {
+pub const MODEL_DEFINITION: AmpModelDefinition = AmpModelDefinition {
     id: MODEL_ID,
     display_name: DISPLAY_NAME,
     brand: BRAND,
-    backend_kind: FullRigBackendKind::Nam,
+    backend_kind: AmpBackendKind::Nam,
     schema,
     validate: validate_params,
     asset_summary,
