+++ crates/block-amp/src/nam_peavey_5150_mesa_4x12.rs
@@ -1,6 +1,6 @@
 use anyhow::{anyhow, Result};
-use crate::registry::FullRigModelDefinition;
-use crate::FullRigBackendKind;
+use crate::registry::{AmpBackendKind, AmpModelDefinition};
+
 use nam::{
     build_processor_with_assets_for_layout, model_schema_for,
     processor::{NamPluginParams, DEFAULT_PLUGIN_PARAMS},
@@ -36,7 +36,7 @@ pub const CAPTURES: &[Peavey5150Capture] = &[
 ];
 
 pub fn model_schema() -> ModelParameterSchema {
-    let mut schema = model_schema_for("full_rig", MODEL_ID, DISPLAY_NAME, false);
+    let mut schema = model_schema_for("amp", MODEL_ID, DISPLAY_NAME, false);
     schema.parameters = vec![
         enum_parameter(
             "boost",
@@ -101,7 +101,7 @@ fn resolve_capture(params: &ParameterSet) -> Result<&'static Peavey5150Capture> 
     }
 
     Err(anyhow!(
-        "full-rig model '{}' does not support boost='{}' mic='{}'",
+        "amp model '{}' does not support boost='{}' mic='{}'",
         MODEL_ID,
         boost,
         mic
@@ -123,11 +123,11 @@ fn build(params: &ParameterSet, sample_rate: f32, layout: AudioChannelLayout) ->
     build_processor_for_model(params, sample_rate, layout)
 }
 
-pub const MODEL_DEFINITION: FullRigModelDefinition = FullRigModelDefinition {
+pub const MODEL_DEFINITION: AmpModelDefinition = AmpModelDefinition {
     id: MODEL_ID,
     display_name: DISPLAY_NAME,
     brand: BRAND,
-    backend_kind: FullRigBackendKind::Nam,
+    backend_kind: AmpBackendKind::Nam,
     schema,
     validate: validate_params,
     asset_summary,
