+++ crates/block-amp/src/nam_vox_ac30.rs
@@ -1,6 +1,6 @@
 use anyhow::{anyhow, Result};
-use crate::registry::FullRigModelDefinition;
-use crate::FullRigBackendKind;
+use crate::registry::{AmpBackendKind, AmpModelDefinition};
+
 use nam::{
     build_processor_with_assets_for_layout, model_schema_for,
     processor::{NamPluginParams, DEFAULT_PLUGIN_PARAMS},
@@ -8,27 +8,25 @@ use nam::{
 use block_core::param::{enum_parameter, required_string, ModelParameterSchema, ParameterSet};
 use block_core::{AudioChannelLayout, BlockProcessor};
 
-pub const MODEL_ID: &str = "synergy_drect_mesa";
-pub const DISPLAY_NAME: &str = "DRECT Mesa";
-const BRAND: &str = "synergy";
+pub const MODEL_ID: &str = "vox_ac30";
+pub const DISPLAY_NAME: &str = "AC30";
+const BRAND: &str = "vox";
 
 pub const NAM_PLUGIN_FIXED_PARAMS: NamPluginParams = DEFAULT_PLUGIN_PARAMS;
 
-const CAPTURE_UNBOOSTED: &str = "full_rigs/synergy_drect_mesa/synergy_drect_unboosted.nam";
-const CAPTURE_OD808_SM57: &str = "full_rigs/synergy_drect_mesa/synergy_drect_od808_sm57.nam";
-const CAPTURE_SD1_SM58: &str = "full_rigs/synergy_drect_mesa/synergy_drect_sd1_sm58.nam";
+const CAPTURE_STANDARD: &str = "full_rigs/vox_ac30/vox_ac30_cab.nam";
+const CAPTURE_CLEAN_65PRINCE: &str = "full_rigs/vox_ac30/vox_ac30_clean_65prince.nam";
 
 pub fn model_schema() -> ModelParameterSchema {
-    let mut schema = model_schema_for("full_rig", MODEL_ID, DISPLAY_NAME, false);
+    let mut schema = model_schema_for("amp", MODEL_ID, DISPLAY_NAME, false);
     schema.parameters = vec![enum_parameter(
-        "boost",
-        "Boost",
+        "character",
+        "Character",
         Some("Amp"),
-        Some("unboosted"),
+        Some("standard"),
         &[
-            ("unboosted", "Unboosted"),
-            ("od808", "OD808"),
-            ("sd1", "SD-1"),
+            ("standard", "Standard"),
+            ("clean_65prince", "Clean 65Prince"),
         ],
     )];
     schema
@@ -59,13 +57,12 @@ pub fn asset_summary(params: &ParameterSet) -> Result<String> {
 }
 
 fn resolve_capture_path(params: &ParameterSet) -> Result<&'static str> {
-    let boost = required_string(params, "boost").map_err(anyhow::Error::msg)?;
-    match boost.as_str() {
-        "unboosted" => Ok(CAPTURE_UNBOOSTED),
-        "od808" => Ok(CAPTURE_OD808_SM57),
-        "sd1" => Ok(CAPTURE_SD1_SM58),
+    let character = required_string(params, "character").map_err(anyhow::Error::msg)?;
+    match character.as_str() {
+        "standard" => Ok(CAPTURE_STANDARD),
+        "clean_65prince" => Ok(CAPTURE_CLEAN_65PRINCE),
         other => Err(anyhow!(
-            "full-rig model '{}' does not support boost='{}'",
+            "amp model '{}' does not support character='{}'",
             MODEL_ID,
             other
         )),
@@ -80,15 +77,15 @@ fn build(params: &ParameterSet, sample_rate: f32, layout: AudioChannelLayout) ->
     build_processor_for_model(params, sample_rate, layout)
 }
 
-pub const MODEL_DEFINITION: FullRigModelDefinition = FullRigModelDefinition {
+pub const MODEL_DEFINITION: AmpModelDefinition = AmpModelDefinition {
     id: MODEL_ID,
     display_name: DISPLAY_NAME,
     brand: BRAND,
-    backend_kind: FullRigBackendKind::Nam,
+    backend_kind: AmpBackendKind::Nam,
     schema,
     validate: validate_params,
     asset_summary,
     build,
-    supported_instruments: block_core::GUITAR_BASS,
+    supported_instruments: block_core::GUITAR_ACOUSTIC_BASS,
     knob_layout: &[],
 };
