+++ crates/block-amp/src/nam_vox_ac30_1961_fawn_ef86.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
-use crate::registry::FullRigModelDefinition;
-use crate::FullRigBackendKind;
+use crate::registry::{AmpBackendKind, AmpModelDefinition};
+
 use nam::{
     build_processor_with_assets_for_layout, model_schema_for,
     processor::{NamPluginParams, DEFAULT_PLUGIN_PARAMS},
@@ -8,16 +8,16 @@ use nam::{
 use block_core::param::{ModelParameterSchema, ParameterSet};
 use block_core::{AudioChannelLayout, BlockProcessor};
 
-pub const MODEL_ID: &str = "marshall_super_100_1966";
-pub const DISPLAY_NAME: &str = "Super 100 1966";
-const BRAND: &str = "marshall";
+pub const MODEL_ID: &str = "vox_ac30_1961_fawn_ef86";
+pub const DISPLAY_NAME: &str = "AC30 '61 Fawn EF86";
+const BRAND: &str = "vox";
 
 pub const NAM_PLUGIN_FIXED_PARAMS: NamPluginParams = DEFAULT_PLUGIN_PARAMS;
 
-const CAPTURE_PATH: &str = "full_rigs/marshall_super_100_1966/marshall_sa100_i_edge_bal_cab.nam";
+const CAPTURE_PATH: &str = "full_rigs/vox_ac30_1961_fawn_ef86/vox_ac30_1961_fawn_crunch.nam";
 
 pub fn model_schema() -> ModelParameterSchema {
-    model_schema_for("full_rig", MODEL_ID, DISPLAY_NAME, false)
+    model_schema_for("amp", MODEL_ID, DISPLAY_NAME, false)
 }
 
 pub fn build_processor_for_model(
@@ -50,15 +50,15 @@ fn build(params: &ParameterSet, sample_rate: f32, layout: AudioChannelLayout) ->
     build_processor_for_model(params, sample_rate, layout)
 }
 
-pub const MODEL_DEFINITION: FullRigModelDefinition = FullRigModelDefinition {
+pub const MODEL_DEFINITION: AmpModelDefinition = AmpModelDefinition {
     id: MODEL_ID,
     display_name: DISPLAY_NAME,
     brand: BRAND,
-    backend_kind: FullRigBackendKind::Nam,
+    backend_kind: AmpBackendKind::Nam,
     schema,
     validate: validate_params,
     asset_summary,
     build,
-    supported_instruments: block_core::GUITAR_BASS,
+    supported_instruments: block_core::GUITAR_ACOUSTIC_BASS,
     knob_layout: &[],
 };
